Automatizacion/requirements.txt
bleach==6.2.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
idna==3.10
kaggle==1.7.4.2
numpy==2.2.4
pandas==2.2.3
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
requests==2.32.3
rsa==4.9
setuptools==77.0.3
six==1.17.0
SQLAlchemy==2.0.39
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
